Easily represent a firm-firm (FF) network using the package network

Create an object of class network from the package network using a FF matrix of class financial_matrix using all the default aesthetic options

Usage

FF.net(x, aesthetic = c("simple", "nice"))

Arguments

x

A matrix-like object produced by FF and related functions.

aesthetic

Choose a pre-set for the network's look. Either 'simple' or 'nice' (see Details).

Value

A network in the desired format

Details

This function does not allow for any of the additional arguments that can be passed to FF.net.custom.

Loops and values

Loops will be allowed if at least one of the matrix's diagonal entries is not zero. The network will be valued if at least one entry of the matrix is neither zero nor one.

Instead, if aesthetic is set to 'simple':

The width of the ties is 1;

The colour of the ties is #b4b4b4 (Philippine Silver);

The size of the nodes is 5;

The colour of the nodes is #081677 (Gentian blue).

Otherwise, if aesthetic is set to 'nice':

The width of the ties is 1;

The colour of the ties is a grey scale reflecting tie strength if the network is valued, otherwise it is #b4b4b4 (Philippine Silver);

The size of the nodes reflects their capitalisation if all firms have data on it and ranges between 1 and 5, otherwise it is 5 for all nodes;

The colour of the nodes reflects their sector if all firms have data on it is taken from a built-in palette, otherwise it is #081677 (Gentian blue).

See also

Author

Telarico, Fabio Ashtar

Examples

# Create a nice network representation of the binary FF of
# Berkshire Hataway's holdings based on common ownership
data("firms_BKB")
x <- FF.naive.ownership(firms_BKB)
FF.net(x = x, aesthetic = 'nice')
#>  Network attributes:
#>   vertices = 47 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 1790 
#>     missing edges= 0 
#>     non-missing edges= 1790 
#> 
#>  Vertex attribute names: 
#>     color size vertex.names 
#> 
#>  Edge attribute names not shown 
#> 
#> 
#> A network representing common ownership 
#> 
#> The size of the nodes represents the firms capitalisation
#> The colour (greyscale) of the ties represents their strength