Create a normalised-valued firm-firm (FF) matrix for boards interlocks

Function to create a normalised-valued firm-firm (FF) matrix based on boards interlocks

Usage

FF.norm.management(
  ...,
  id_as_firm_name = NULL,
  Matrix = NULL,
  self_ties = FALSE
)

Arguments

...

Either multiple objects of class firm or a list of such objects

id_as_firm_name

Whether to use the ticker as the firm's name. Defaults to TRUE if all firms' id is neither NULL nor NA.

Matrix

Whether to use the Matrix package. Defaults to TRUE when any matrix in the pipeline contains more than 10,000 cells and the package is installed.

self_ties

Whether to allow self-ties (a 'loop' in graph theory). Defaults to FALSE.

Value

A matrix object of class financial_matrix(possibly using the Matrix package)

Details

Normalised-valued means that weights represent the share of common managers.

See also

Author

Telarico, Fabio Ashtar

Examples

# Create the normalised FF matrix of Berkshire Hathaway's holdings by boards interlocks
data('firms_BKB')
FF <- FF.norm.management(firms_BKB)