Calculate the Fiedler value (algebraic connectivity)

This function expresses the algebraic connectivity of a Firm-Firm network as its Fiedler value. The Fiedler value, named after Miroslav Fiedler, who explored its significance, summarises the connectivity and robustness of a network. Mathematically, it is the second smallest eigenvalue of the network's Laplacian matrix

Usage

fiedler(..., ignore.weights = FALSE, generalise = NULL)

Arguments

...

Firm-Firm network in one of the following classes:

financial_matrix produced by FF and family;

network_financial or network if the relevant package is installed;

igraph_financial or igraph if the relevant package is installed.

ignore.weights

Optional parameter, defaults to FALSE. If TRUE, ignore ties weights in the computation.

generalise

See Details for more information. Defaults to NULL.

Value

A numeric, the Fiedler value.

Details

The Fiedler value is specifically defined for undirected graphs. For directed or asymmetrical networks like the Firm-Firm ones, the Laplacian is not necessarily symmetric, and its eigenvalues can be complex. In practical applications, this is more likely to happen in naively valued networks, due to the presence of large off-diagonal values. But it can happen also under other circumstances.

There are three main workarounds:

1. Symmetrisation - Simply considers the underlying undirected graph. This involves ignoring the direction of edges and calculating the Laplacian matrix and its eigenvalues as if the graph were undirected.

2. The generalised Laplacian calculated as the sum of the diagonal matrices for in-degree and out-degree \(\left([D]_{in}+[D]_{out}\right)/2\)

3. The Hermitian part of the Laplacian - Uses the Hermitian part of the Laplacian matrix of the directed network \(\left([M]+{[M]^\text{C}}^\top)\right)/2\), where the second addendum is the conjugate transpose of the adjacency matrix.

Practically, the third method is excessive here, as the values of the ties cannot be complex numbers. Indeed, the Hermitian is actually the Laplacian of the underlying symmetric network with the value of the ties being split equally in both directions because the conjugate of a real number is that number. Moreover, symmetrising before calculating the Laplacian or generalising the matrix \([L]\) returns the same result. So, the parameter generalise is logical and takes the following values:

TRUE for the generalised Laplacian;

FALSE for the possibly complex (and uninterpretable) eigenvalue of the as-is Laplacian.

NULL will take the generalised Laplacian only if necessary.

References

Fiedler, Miroslav. ‘Laplacian of Graphs and Algebraic Connectivity’. Banach Center Publications 25, no. 1 (1989): 57–70. https://eudml.org/doc/267812.

Guo, Krystal, and Bojan Mohar. ‘Hermitian Adjacency Matrix of Digraphs and Mixed Graphs’. Journal of Graph Theory 85, no. 1 (May 2017): 217–48. doi:10.1002/jgt.22057 .

Author

Telarico, Fabio Ashtar

Examples

# Load some data
data('firms_BKB')
# Create a FF matrix
mat <- FF(firms_BKB, who = 'b', ties =  'n')
fiedler(mat)
#> [1] 217.7835

# Create a FF network
if(!require('network')){
  net <- FF.net(mat, 'simple')
  fiedler(net)==fiedler(mat)
}
#> Loading required package: network
#> 
#> ‘network’ 1.18.2 (2023-12-04), part of the Statnet Project
#> * ‘news(package="network")’ for changes since last version
#> * ‘citation("network")’ for citation information
#> * ‘https://statnet.org’ for help, support, and other information
#> 
#> Attaching package: ‘network’
#> The following object is masked from ‘package:FinNet’:
#> 
#>     network.size

# Create a FF graph
if(!require('igraph')){
  g <- FF.graph(mat, 'simple')
  fiedler(g)==fiedler(mat)
}
#> Loading required package: igraph
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:network’:
#> 
#>     %c%, %s%, add.edges, add.vertices, delete.edges, delete.vertices,
#>     get.edge.attribute, get.edges, get.vertex.attribute, is.bipartite,
#>     is.directed, list.edge.attributes, list.vertex.attributes,
#>     set.edge.attribute, set.vertex.attribute
#> The following objects are masked from ‘package:FinNet’:
#> 
#>     E, V, ecount, gorder, gsize, vcount
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union