Function to create a firm (legal person) using data from 'Yahoo! Finance'

Tickers can be retrieved from [Yahoo! Finance](https://finance.yahoo.com/lookup/). This function requires the package yahoofinancer to be installed. It is available from the CRAN by running install.packages('yahoofinancer').

Usage

find.firm(
  ticker,
  name = NULL,
  ticker_is_id = TRUE,
  legal_form = NULL,
  sector_granularity = 1,
  managers_remove_salutation_title = TRUE,
  managers_only_surname = FALSE
)

Arguments

ticker

Firm's ticker.

name

Provide the firm's name. If not provided, NA, or NULL, will default to the string provided as ticker.

ticker_is_id

Should the ticker be used as the firm's id?

legal_form

The firm's legal form of the firm. Possible values: - a string (e.g., 'LLC', 'Private', 'GmbH', etc.); - NULL (default), in which case the function will set legal_form to 'JSC'; or - NA to specify no legal form.

sector_granularity

Sector in which the firm operates. Possible values: - 0, NULL, or NA to omit the sector; - 1 or 'generic' (default) for a generic description (e.g., 'Consumer Technology', 'Consumer Cyclical', 'Consumer Defensive'); - 2 or 'specifc' for a more granular description (e.g., 'Technology', 'Auto Manufacturers', 'Tobacco').

managers_remove_salutation_title

Yahoo! Finance provide salutation titles before the names of the managers. If this is TRUE (default), they will be omitted.

managers_only_surname

Yahoo! Finance provide first, middle, and last name of the managers. If this is TRUE (not recommended for large data sets), only the surname is returned.

Value

An object of the S4 class firm containing several fields, only the first one of which is mandatory:

name

Name of the firm (or ticker if no name was provided)

id

Firm' ticker (if ticker_is_id was `TRUE`) or nothing (otherwise)

legal_form

Legal form of the firm (may be null)

sector

Sector in which the firm operates (may be null)

revenues

Yearly revenues

capitalisation

Capitalisation

management

Members of the board

ownership

Owner(s)

shares

Share owned by (each of) the owner(s)

currency

Currency

See also

Author

Telarico, Fabio Ashtar

Examples

# Registering Apple automatically
AAPL <- find.firm(ticker = 'AAPL', name = 'Apple', legal_form = 'Inc')