Extending igraph functions to igraph_financial objects

The following functions are implemented:

V_fin to retrieve the vertexes (igraph::V);

vcount_fin to count the vertexes (igraph::vcount);

gorder_fin as an alias to vcount_fin (igraph::gorder);

E_fin to retrieve the edges (igraph::E);

gsize_fin to count the edges (igraph::gsize);

ecount_fin as an alias to gsize_fin (igraph::ecount)

plot_igraph_fin to plot graphs (igraph::plot.igraph))

Usage

V(x)

vcount(x)

gorder(x)

E(x, ...)

ecount(x, ...)

gsize(x, ...)

plot_igraph(x, ...)

Arguments

x

The igraph_financial object

...

Other parameters passed to the corresponding igraph functions (see Details).

Value

The same result for both igraph and igraph_financial objects

V: A vertex sequence containing all vertices, in the order of their numeric vertex ids.

vcount and gorder: Number of vertices, numeric scalar.

E: An edge sequence of the graph

ecount and gsize: Number of edges, numeric scalar.

plot_igraph: Returns NULL, invisibly. Called to print the graph to any R device. (see method and igraph::plot.igraph)

Details

Implementing most basic iterators from the package igraph for objects of class igraph_financial

Author

Telarico, Fabio Ashtar